Lager midlertidig utbetlaling-komponent
