+++ src/app/personside/infotabs/utbetalinger/Utbetalinger.tsx
@@ -8,8 +8,10 @@ import { Undertekst, Undertittel } from 'nav-frontend-typografi';
 import { restoreScroll } from '../../../../utils/restoreScroll';
 import { FilterState, default as Filtrering } from './Filter';
 import TotaltUtbetalt from './TotaltUtbetalt';
-import { AlignTextCenter } from '../../../../components/common-styled-components';
-import UtbetalingerListe from './UtbetalingerListe';
+import { AlignTextCenter, Bold, Uppercase } from '../../../../components/common-styled-components';
+import { ArrayGroup, groupArray, GroupedArray } from '../../../../utils/groupArray';
+import { månedOgÅrForUtbetaling, sortByPosteringsDato } from './utbetalingerUtils';
+import EnkeltUtbetaling from './Utbetaling';
 
 export interface UtbetalingerProps {
     utbetalinger: Utbetaling[];
@@ -51,6 +53,19 @@ const ArenaLenkeStyle = styled.div`
   text-align: right;
 `;
 
+const MånedGruppeStyle = styled.div`
+  > *:first-child {
+    background-color: ${theme.color.bakgrunn};
+    padding: .2rem 1.2rem;
+  }
+`;
+
+const UtbetalingerStyle = styled.div`
+  > *:first-child {
+    padding: 0 1.2rem .2rem;
+  }
+`;
+
 function ArenaLenke() {
     return (
         <ArenaLenkeStyle>
@@ -58,6 +73,15 @@ function ArenaLenke() {
                 Lenke til Arena: <a className="lenke">Meldinger/utbetalinger i Arena</a>
             </Undertekst>
         </ArenaLenkeStyle>
+    );
+}
+function Månedsgruppe({gruppe}: { gruppe: ArrayGroup<Utbetaling> }) {
+    return (
+        <MånedGruppeStyle>
+            <Undertekst><Bold><Uppercase>{gruppe.category}</Uppercase></Bold></Undertekst>
+            {gruppe.array.map(utbetaling =>
+                <EnkeltUtbetaling utbetaling={utbetaling} key={utbetaling.posteringsdato}/>)}
+        </MånedGruppeStyle>
     );
 }
 
@@ -71,6 +95,11 @@ function Utbetalinger(props: UtbetalingerProps) {
         );
     }
 
+    const utbetalingerGruppert: GroupedArray<Utbetaling> = groupArray(
+        props.utbetalinger.sort(sortByPosteringsDato),
+        månedOgÅrForUtbetaling
+    );
+
     return (
         <Wrapper>
             <Venstre onClick={restoreScroll}>
@@ -80,7 +109,12 @@ function Utbetalinger(props: UtbetalingerProps) {
             <Hoyre>
                 <ArenaLenke/>
                 <TotaltUtbetalt utbetalinger={props.utbetalinger} filter={props.filter}/>
-                <UtbetalingerListe utbetalinger={props.utbetalinger}/>
+                <UtbetalingerStyle>
+                    <Undertittel>Utbetalinger</Undertittel>
+                    {utbetalingerGruppert.map((gruppe: ArrayGroup<Utbetaling>) =>
+                            <Månedsgruppe gruppe={gruppe} key={gruppe.category} />
+                    )}
+                </UtbetalingerStyle>
             </Hoyre>
         </Wrapper>
     );
